requirements.txt
aws-lambda-powertools==1.20.1
chalice==1.24.2
psycopg2-binary==2.9.1